Upgrading dependencies, moving files to the correct folder and removing momentjs
