+++ package.json
@@ -38,26 +38,21 @@
     "not op_mini all"
   ],
   "dependencies": {
-    "bootstrap": "4.3.1",
+    "bootstrap": "^4.6.0",
     "classnames": "2.2.6",
-    "headroom.js": "0.11.0",
-    "moment": "2.24.0",
-    "node-sass": "4.13.1",
-    "nouislider": "14.1.1",
-    "react": "16.12.0",
-    "react-datetime": "2.16.3",
-    "react-dom": "16.12.0",
-    "react-router": "5.1.2",
-    "react-router-dom": "5.1.2",
-    "react-scripts": "^3.4.1",
-    "reactstrap": "8.4.1"
+    "headroom.js": "0.12.0",
+    "node-sass": "5.0.0",
+    "nouislider": "14.6.3",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "react-router": "5.2.0",
+    "react-router-dom": "5.2.0",
+    "react-scripts": "^4.0.3",
+    "reactstrap": "8.9.0"
   },
   "devDependencies": {
-    "@types/googlemaps": "3.39.2",
-    "@types/markerclustererplus": "2.1.33",
-    "@types/react": "16.9.19",
-    "eslint-plugin-flowtype": "3.13.0",
-    "jquery": "3.4.1",
-    "typescript": "3.7.5"
+    "@types/react": "17.0.3",
+    "eslint-plugin-flowtype": "5.4.0",
+    "jquery": "3.6.0"
   }
 }
