+++ src/assets/scss/argon-design-system/vendors.scss
@@ -1,5 +1,4 @@
 
-@import "vendor/datetimepicker";
 @import "vendor/_bootstrap-datepicker";
 @import "vendor/nouislider";
