+++ src/assets/scss/react/react-differences.scss
@@ -8,5 +8,3 @@
 
 
 // Vendor
-
-@import "custom/vendor/react-datetime.scss";
