+++ src/index.js
@@ -6,22 +6,28 @@ import "assets/vendor/nucleo/css/nucleo.css";
 import "assets/vendor/font-awesome/css/font-awesome.min.css";
 import "assets/scss/argon-design-system-react.scss?v1.1.0";
 
-import Index from "views/Index.js";
-import Landing from "views/examples/Landing.js";
-import Login from "views/examples/Login.js";
-import Profile from "views/examples/Profile.js";
-import Register from "views/examples/Register.js";
+import Index from "./views/Index.js";
+import Landing from "./views/examples/Landing.js";
+import Login from "./views/Login/Login.js";
+import Profile from "./views/examples/Profile.js";
+import Register from "./views/Register/Register.js";
 
 ReactDOM.render(
   <BrowserRouter>
     <Switch>
-      <Route path="/" exact render={props => <Index {...props} />} />
+      <Route path="/"
+        exact
+        render={props => <Index {...props} />}
+      />
       <Route
         path="/landing-page"
         exact
         render={props => <Landing {...props} />}
       />
-      <Route path="/login-page" exact render={props => <Login {...props} />} />
+      <Route path="/login-page"
+        exact
+        render={props => <Login {...props} />}
+      />
       <Route
         path="/profile-page"
         exact
